Resend only old packets
Do not resend packets that got published within the last retry_interval.

Before this commit, it can happen that a packet is resent even if the retry
interval is 30 seconds and a packet gets acked by the server within a few
milliseconds.

Also, retry_interval unit was wrong:
The library user defines retry_interval in seconds.
Within the library, retry_interval is milliseconds.