fix(ssl): correctly handle ssl error events
`emqx_sock` defined an internal record for storing both the tcp and
ssl socket ports.  However, when we have a TLS error such as
`bad_certificate`, `ssl` sends a message containing the inner SSL
socket, not this internal emqtt record, and there were missing clauses
to handle that.

This caused `emqtt` to:
i) Hang indefinitely if `connect_timeout` is not specified;
ii) Fail with `{error, connack_timeout}` if `connect_timeout` is a finite value.