Support multiple subscription identifiers
The MQTT 5.0 spec states
```
The Subscription Identifier can have the value of 1 to 268,435,455.
It is a Protocol Error if the Subscription Identifier has a value of 0.
Multiple Subscription Identifiers will be included if the publication
is the result of a match to more than one subscription, in this case
their order is not significant.
[...]
If the Client specified a Subscription Identifier for any of the
overlapping subscriptions the Server MUST send those Subscription
Identifiers in the message which is published as the result of the
subscriptions [MQTT-3.3.4-3]. If the Server sends a single copy of
the message it MUST include in the PUBLISH packet the Subscription
Identifiers for all matching subscriptions which have a Subscription
Identifiers, their order is not significant [MQTT-3.3.4-4]. If the
Server sends multiple PUBLISH packets it MUST send, in each of them,
the Subscription Identifier of the matching subscription if it has a
Subscription Identifier [MQTT-3.3.4-5].
```

This commit allows parsing multiple subscription identifiers if the
server sends a single copy.

Technically, this client library only needs to parse multiple subscription
identifiers. This is because the server can send multiple subscription
identifiers to the client in the PUBLISH packet. However, the client
is not allowed to send multiple subscription identifiers to the server
in the SUBSCRIBE packet (and none at all in the PUBLISH packet).
However, in order for the symmetric property test to succeed, this
commit also adds the serialisation part.

Further, to keep this library backwards compatible, if the server sends
a single subscription identifier, the client app will continue to receive an
integer. The client app receives a list only if the server sends multiple
subscription identifiers in the PUBLISH packet.

This commit also fixes a wrong type sepc in the README.