refactor: avoid crashing when CONNECT packet send error
Prior to this fix, if the CONNECT packet is sent to an already
closed socket, the process would crash because the 'stop_and_reply'
reason is not a {shutdown, Reason} tuple, but just 'Reason'.
e.g. 'closed'

This commit changed the code to handle all errors happened
during the connection establishment and the first CONNECT
packet send as {shutdown, Reason} so there would be no crash logs.

Also, a debug level log is added for each send error.